func_80035E20: hold the tpage byte in a u32 (6 register rows to 5)
The tpage `or` came out `or $v0,$v0,$a0` where retail has
`or $v0,$a0,$v0`. Seven hand spellings of the expression tied. A permuter
run from the previous head found that the width of `b` decides the
operand order: with `u32 b` the `or` matches.

`b` is only assigned the byte at +0x66, has 0x80 or-ed in, and is or-ed
into the halfword tpage, so the wider local stores the same value.

Every opcode position still matches. Five rows remain, all in one place:
after GsSetLsMatrix, retail computes the three RotAverageNclip4 vector
addresses before materialising -8 and 8.

+++ src/candidates/func_80035E20.c
@@ -11,12 +11,13 @@
  *   the flag test, which is where retail has it, and they also let the
  *   scheduler split the 0x1F8000A0 constant's lui and ori the way retail
  *   does;
+ * - `b` is a u32: its width decides the operand order of the tpage `or`,
+ *   where seven spellings of the expression itself tied;
  * - the -fno-strength-reduce profile is the measured one: plain
  *   gcc_2_8_1_g8_split is 886 instructions.
  *
- * Residual: every opcode position matches; six rows differ in registers or
- * operand order. The tpage `or` takes its operands in the other order, and
- * after GsSetLsMatrix the target computes the three RotAverageNclip4 vector
+ * Residual: every opcode position matches; five rows differ in registers.
+ * After GsSetLsMatrix the target computes the three RotAverageNclip4 vector
  * addresses before materialising -8 and 8, where this source materialises the
  * constants first.
  */
@@ -51,7 +52,7 @@ void func_80035E20(DisplayObject *obj, GsOT *ot)
     s32 cyv;
     s32 c;
     u8 f;
-    u8 b;
+    u32 b;
     s32 v78;
     s32 tt;
     s16 k8;
